Split StatefulMafsGraph and MafsGraph into separate files
